feat: added empty search result view
