+++ static/js/src/store/components/App/App.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import { QueryClient, QueryClientProvider } from "react-query";
 
-import Packages from "../Packages";
+import Packages from "../../pages/Packages";
 
 function App() {
   const queryClient = new QueryClient({
