+++ static/js/src/store/components/App/__tests__/App.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from "@testing-library/react";
 import App from "../App";
 import "@testing-library/jest-dom";
 
-jest.mock("../../Packages", () => () => <div>Packages Component</div>);
+jest.mock("../../../pages/Packages", () => () => <div>Packages Component</div>);
 
 describe("App Component", () => {
   test("should render the App component with the Packages route", () => {
