+++ static/js/src/store/components/EmptyResultSection/index.ts
@@ -1,0 +1,1 @@
+export { EmptyResultSection } from "./EmptyResultSection";
